requirements.txt
Django>=3.2.0,<3.3.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
drf_yasg==1.21.7
psycopg2-binary>=2.9.1,<3.0.0
gunicorn==20.1.0
whitenoise==5.3.0
